requirements.txt
vispy>=0.14.2
numpy>=1.24
Pillow>=10.0
PyOpenGL>=3.1.7
